test(authentication/services unit): refactored the unit tests
Refactored the unit tests and divided into authentication and services, to reflect the new project
naming convention.
